Fix bug on new form
There was a bug when trying to go to new form page on the home page. When using the Status "waiting", "approved" and "delivered" it had the status showing in the url, then when clicking on the button it was trying to open "home/order/form" which doesn't exist. The solution was create a new method in the "homeController" to redirect the request to "/order/form".